+++ src/saedb/aggregator/iaggregator.hpp
