+++ src/saedb/sae_include.hpp
@@ -2,7 +2,7 @@
 #define SAE_INCLUDE_ALL
 
 #include "graph.hpp"
-#include "engine.hpp"
+#include "iengine.hpp"
 #include "ialgorithm.hpp"
 #include "synchronous_engine.hpp"
